Use racket/base instead of racket
```#lang racket``` pulls in a lot of extra libraries that aren't necessary for Hello World, which slows down the program quite a bit.

I'm posting this suggestion in response to a conversation I had with someone who was complaining Racket is slow on such a trivial program. I believe most of us who write scripts in Racket will usually use ```racket/base``` unless we *need* to bring in many of those other libraries.